+++ client/gui/GUITextBox.hpp
@@ -85,12 +85,12 @@ public:
 
   virtual void draw(double elapseTime)
   {
-    this->_sprite.draw(this->_x, this->_y, elapseTime);
+    this->_sprite.draw(static_cast<int>(this->_x), static_cast<int>(this->_y), elapseTime);
     if (this->_font)
       {
 	this->_font->setText(this->_text);
-	this->_font->draw(this->_x + ((this->_sprite.getWidth() - this->_font->getWidth()) / 2),
-			  this->_y + ((this->_sprite.getHeight() - this->_font->getHeight()) / 2) - 5, elapseTime);
+	this->_font->draw(static_cast<int>(this->_x + ((this->_sprite.getWidth() - this->_font->getWidth()) / 2)),
+			  static_cast<int>(this->_y + ((this->_sprite.getHeight() - this->_font->getHeight()) / 2) - 5), elapseTime);
       }
   }
